third_party/blink/web_tests/external/wpt/css/css-page/parsing/margin-rules-001-expected.txt

This is a testharness.js-based test.
[FAIL] @page { @top-left-corner { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @top-left-corner { } }" but got "@page { }"
[FAIL] @page { @top-left { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @top-left { } }" but got "@page { }"
[FAIL] @page { @top-center { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @top-center { } }" but got "@page { }"
[FAIL] @page { @top-right { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @top-right { } }" but got "@page { }"
[FAIL] @page { @top-right-corner { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @top-right-corner { } }" but got "@page { }"
[FAIL] @page { @bottom-left-corner { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @bottom-left-corner { } }" but got "@page { }"
[FAIL] @page { @bottom-left { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @bottom-left { } }" but got "@page { }"
[FAIL] @page { @bottom-center { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @bottom-center { } }" but got "@page { }"
[FAIL] @page { @bottom-right { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @bottom-right { } }" but got "@page { }"
[FAIL] @page { @bottom-right-corner { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @bottom-right-corner { } }" but got "@page { }"
[FAIL] @page { @left-top { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @left-top { } }" but got "@page { }"
[FAIL] @page { @left-middle { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @left-middle { } }" but got "@page { }"
[FAIL] @page { @left-bottom { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @left-bottom { } }" but got "@page { }"
[FAIL] @page { @right-top { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @right-top { } }" but got "@page { }"
[FAIL] @page { @right-middle { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @right-middle { } }" but got "@page { }"
[FAIL] @page { @right-bottom { } } should be a valid rule
  assert_equals: serialization should be canonical expected "@page { @right-bottom { } }" but got "@page { }"
Harness: the test ran to completion.
